input, output order

+++ oneflow-backend/src/model_state.h
@@ -56,7 +56,7 @@ namespace triton { namespace backend { namespace oneflow {
 struct InputOutputAttribute {
   TRITONSERVER_DataType datatype;
   std::vector<int64_t> input_shape;
-  int input_index;
+  size_t input_index;
 };
 
 //
